% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad}
\alias{pad}
\title{Point Anomaly Detection}
\usage{
pad(ts, method, ...)
}
\arguments{
\item{ts}{An univariate time series}

\item{method}{Name of the desired method}

\item{...}{Parameters for the specified method (see specific help for more details)}
}
\value{
list of class \code{pad}, including two vectors :
  \item{\code{ts}}{original time series}
  \item{\code{anomalies}}{estimated anomalies (indexes)}
}
\description{
The \code{pad} function propose a wide range of methods to detect 
point anomalies in a univariate time-serie.
}

