% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.sw.median.R
\name{pad.sw.median}
\alias{pad.sw.median}
\title{Point anomaly detection using sliding window median}
\usage{
pad.sw.median(ts, side = "one-sided", size = 10, threshold = 1)
}
\arguments{
\item{ts}{An univariate time series}

\item{side}{Choose between \code{one-sided} or \code{two-sided}}

\item{size}{Size of the sliding window}

\item{threshold}{}
}
\value{
Vector of estimated anomalies (indexes)
}
\description{
a value $t_i$ is considered as an anomaly if \eqn{abs(t_i - m_i) > threshold}, where \eqn{m_i}
is the median of the neighborhood of \eqn{i} defined with the following (where \eqn{k = size}):
\describe{
  \item{\code{two-sided}}{\eqn{m_i = median(t_(i-k), ..., t_(i+k))}, then \eqn{t_i} is inside}
  \item{\code{one-sided}}{\eqn{m_i = median(t_(i-2k), ..., t_(i-1))}, then \eqn{t_i} is outside}
}
}
\references{
% bibentry: Basu2007
Basu S and Meckesheimer M (2007).
``Automatic outlier detection for time series: an application to sensor data.''
\emph{Knowledge and Information Systems}, \bold{11}(2), pp. 137--154.
ISSN 0219-1377, \url{http://doi.org/10.1007/s10115-006-0026-6}, \url{http://dx.doi.org/10.1007/s10115-006-0026-6}.
% end:bibentry:  Basu2007
}
\seealso{
\code{\link{pad}}
}

